会员完成

+++ gmall-pms/src/main/java/com/atguigu/gmall/pms/mapper/ProductCategoryMapper.java
@@ -16,5 +16,5 @@ import java.util.List;
  */
 public interface ProductCategoryMapper extends BaseMapper<ProductCategory> {
 
-    List<ProductCategoryVo> selectLists(int id);
+    List<ProductCategoryVo> selectLists(Long id);
 }
